fix(model-resolver): single thinking-suffix parser (#2379 follow-up)
Consolidate the two remaining duplicate thinking-suffix parsers onto the
canonical splitSelectorThinkingSuffix:
- move splitSelectorThinkingSuffix + SelectorThinkingSuffix into thinking.ts
  (co-located with parseThinkingLevel, avoiding an import cycle) and
  re-export from config/model-resolver for existing importers;
- thinking.ts formatClampedModelSelector now delegates its id-suffix parse
  to the canonical splitter instead of an inline lastIndexOf/parseThinkingLevel;
- slash-commands/builtin-registry splitExplicitThinkingSelector delegates to
  the canonical splitter, preserving whole-selector fallback on invalid suffix.

The OpenRouter route/variant suffix parser (getOpenRouterRouteSuffix) is a
distinct non-thinking concern and is intentionally left intact.

Refs #2379
